+++ package.json
@@ -6,8 +6,11 @@
     "node": "20.x"
   },
   "scripts": {
-    "dev": "npx prisma generate && next dev",
-    "build": "npx prisma generate && next build",
+    "gen:prisma": "npx prisma generate",
+    "db:pull": "npx prisma db pull",
+    "db:push": "npx prisma db push",
+    "dev": "next dev",
+    "build": "gen:prisma && next build",
     "hardhat": "npx hardhat node",
     "hardhat:fund": "npx hardhat accounts --network custom",
     "start": "next start",
