+++ src/app/providers.tsx
@@ -3,12 +3,11 @@
 import { CacheProvider } from '@chakra-ui/next-js'
 import { ChakraProvider, ColorModeScript } from '@chakra-ui/react'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import { watchAccount } from '@wagmi/core'
 import { createWeb3Modal } from '@web3modal/wagmi/react'
-import { useEffect, useState } from 'react'
 import { WagmiProvider } from 'wagmi'
 
 import { UserProvider } from '@/app/user-provider'
+import { useUpsertUser } from '@/lib/hooks/use-upsert-user'
 import { ProtectionProvider } from '@/lib/providers/protection'
 import { SignTermsProvider } from '@/lib/providers/sign-terms-provider'
 import theme from '@/lib/styles/theme'
@@ -16,8 +15,6 @@ import { config, metadata, projectId } from '@/lib/utils/wagmi'
 
 import '@/lib/styles/fonts'
 import { AnalyticsProvider } from './analytics-provider'
-
-import type { User } from '@prisma/client'
 
 
 const queryClient = new QueryClient()
@@ -34,30 +31,7 @@ createWeb3Modal({
 })
 
 export function Providers({ children }: { children: React.ReactNode }) {
-  const [persistedUserData, setPersistedUserData] = useState<User | null>(null)
-
-  useEffect(() => {
-    const unwatch = watchAccount(config, {
-      async onChange(account) {
-        try {
-          const user = await (
-            await fetch('/api/user', {
-              method: 'POST',
-              body: JSON.stringify({
-                address: account.address,
-              }),
-            })
-          ).json()
-
-          setPersistedUserData(user)
-        } catch (e) {
-          console.error(e)
-        }
-      },
-    })
-
-    return () => unwatch()
-  }, [])
+  const persistentUserData = useUpsertUser()
 
   return (
     <CacheProvider prepend={true}>
@@ -68,7 +42,7 @@ export function Providers({ children }: { children: React.ReactNode }) {
             <AnalyticsProvider>
               <ProtectionProvider>
                 <SignTermsProvider>
-                  <UserProvider value={persistedUserData}>
+                  <UserProvider value={persistentUserData}>
                     {children}
                   </UserProvider>
                 </SignTermsProvider>
